v0.3 Added Zipper
Added zippers below side menus that when user selects suggestMovieButton close.

Note: Pressing suggestMovieButton temporarily disables the clearInputButton that triggers the open animation to prevent starting the open animation when the zipper is sill closing.

Note: When the zipper animation finished the user should be presented a screen with movie suggestions. The logic mentioned above that handles the enabling and disabling of the buttons that trigger the zipper animations should probably be altered:
- On the presented screen holding the movie suggestions there should be an action to dismiss the screen and open the zipper to reset the selection procedure. 

Ideas: Might be nice to change the most inner part of the zipper to match the color of the menu. This could probably be done by introducing a gradient with a hard transition from light grey to the darker menu grey.