v0.6 Reworked (Zipper) Constraints
Zipper constraints have been updated to ensure zipper's (as a whole) center y anchor matches center y anchor of the sideMenu instead of previously being top-anchored to the top anchor of the menu. This results in neater equal spacing for the top and bottom spacing independent of iPhone model. 

TopBar height and suggestMovieButton size are now proportionally instead of being hardcoded to allow scaling between iPhone models.

Test: Check old smaller iPhone models to check wether zipper is contained in side menu and if layout looks decent.

Investigate possibility to set model-dependent values for certain constraints to optimise for most and/or all different models