v0.9 Fixed Dimmer Issue
Context: When suggestMovieButton gets tapped until the open zipper animations completes the modeSelectionButton is disabled to prevent any reset of the animation while running by pressing the modeSelectionButton. However the icon for the modeSelection button did not properly dim anymore. 

While refactoring, addObservers was not removed from the setupButton method resulting in it being added twice and because of that ended up not working. It is not being called only once, during initialisation.