v0.11 MoviesSuggestionsScreen
Started building UI of the movieSuggestionsScreen. If suggestMovieButton gets tapped, and after the zipper closes, the movieSuggestionsScreen is presented to user as an overlay. This screen will be used to display the movie suggestions and allows the option to navigate between the several movie suggestions from best match to alternative matches.

This can be done either by:
- Swiping left or right on the suggestionsScreen 
- By tapping the triangle-shaped navigators on the side of the suggestionsScreen

Todo: Display movie information on suggestionScreen:
- Name of movie
- Image
- Lead role(s)
- Director?
- Length