v0.13 Added Genre Selection Screen
Tapping a button on the left or right selection menu will launch selection screen of which the content will later be used to make a filtered request to TheMovieDB API. 

At the moment a selection screen is presented, populated by genres, when the top button on the left side menu is tapped. The genres in the tableView can be tapped to change their state between false and true. These states will be used to define a maximum of genres that can be selected per user and to define the filter for the api call. 

Eventually, genres selected by each user will be used to construct an algorithm to present movie options. The user(s) is then able to navigate the suggestions by swiping/tapping.