v0.16 Added networking for people
The networking layer, for handling the people data, is finished and ready for testing. It uses a generic networking layer that handles the pages for the people data on which a networking layer is built that handles unwrapping that data into usable Person data. It currently iteration over the pages and is limited by either a total amount of pages and a set maximum of pages (currently 50) that should be loaded. 

Next: Test people networking
Next: Build networking layer test for movies (discover). This should for now handle the fetching of movies with a pre-set genre-id array. This will later be modified to hold a dynamic set of genre-ids which depending on user(s) choice(s).

Next: The third criterium that a user can edit will likely be a post API-call filter (e.g. protagonist sex, movie length)