v0.19 Person Networking
Finished networking for popular person. When launched the second button in the left selectionMenu (currently left-side only) correctly launches a tableView populated with peopleData. It is currently set to fetch 5 pages allowing enough diversity (for now).

The optionsMenus as well as the movieSuggestionsView are now properly altered when light/darkmode is selected.

Fixed issue: When tapping the clearInputButton the selected state change (white labelText) for the optionsMenus would not reset to its default color black. This has been fixed by adding the following to the removeSelection method;

 if let indexes = options.indexPathsForSelectedRows {
      for index in indexes {
             options.deselectRow(at: index, animated: true)
             options.cellForRow(at: index)?.textLabel?.textColor = UIColor.black
       }
       options.reloadData()
 }