v0.20 Refactor, Slider, Reachability
Refactored the left/rightSelectionMenu and moved code to the selectionMenu superclass. Did same for genre/personOptionsMenu for which code has been moved to optionsMenu. These superclasses now hold all or most shared code for the subclasses. Left/right selectionMenu now properly responds to light/darkmode changes by adjusting alpha.

The third selectionMenu contains a slider which allows user(s) to slide-select a minimum vote average of which the value is directly visible in the ratingLabel. The slider-thumb image has been set to a small zipper.

There is now a Reachability object that checks wether there is an internet connection when user is about to enter a selectionMenu. If there is no connection the user will not be taken to the selectionMenu.

Idea: Currently when data for a selectionMenu is loading there is a label visible with "loading genre data...". It would be nice to have an activity indicator instead.

Todo:
- Save/transfer user selections to an object
- Edit selection information to suit API call, representing user preference
- Make API call for discover movies using the (edited) data
- Set up movieSuggestionsView